Improve `CancellationException`
Motivation:

1. `H2ClientParentConnectionContext.StacklessCancellationException` was
reused by `SpliceFlatStreamToMetaSingle`, which created a false
impression in logs that this exception is related to HTTP/2.
2. There is not much value in a stacktrace for `CancellationException`
in `BeforeFinallyHttpOperator`.

Modifications:

1. Make `H2ClientParentConnectionContext.StacklessCancellationException`
a top level class.
2. Create a `StacklessCancellationException` in http-utils module for
`BeforeFinallyHttpOperator`.
3. Add exception message in `DefaultBlockingIterableProcessor`.

Result:

Easier for users to reason about received `CancellationException`.
